refactor: 폴더 분리
용도에 맞게 폴더에 적절히 넣어 분리
