+++ src/main/java/org/example/demo/domain/Post.java
@@ -1,4 +1,4 @@
-package org.example.demo;
+package org.example.demo.domain;
 
 import java.time.LocalDateTime;
