+++ src/main/java/org/example/demo/domain/User.java
@@ -1,4 +1,4 @@
-package org.example.demo;
+package org.example.demo.domain;
 
 public class User {
     private Long id;
