+++ src/main/java/org/example/demo/servlet/PostServlet.java
@@ -1,10 +1,12 @@
-package org.example.demo;
+package org.example.demo.servlet;
 
 import jakarta.servlet.ServletException;
 import jakarta.servlet.annotation.WebServlet;
 import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
+import org.example.demo.db.PostDb;
+import org.example.demo.domain.Post;
 
 import java.io.IOException;
