+++ src/main/java/org/example/demo/servlet/UserUpdateServlet.java
@@ -1,10 +1,11 @@
-package org.example.demo;
+package org.example.demo.servlet;
 
 import jakarta.servlet.ServletException;
-import jakarta.servlet.annotation.WebServlet;
 import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
+import org.example.demo.db.UserDb;
+import org.example.demo.domain.User;
 
 import java.io.IOException;
